requirements.txt
# CV Automation System - Python Dependencies
# Last Updated: October 1, 2025

# ============================================================================
# Core PDF/Document Processing
# ============================================================================
PyPDF2==3.0.1                    # PDF text extraction
python-docx==1.1.0               # DOCX processing and generation
pdf2image==1.16.3                # PDF to image conversion (for OCR)
Pillow==10.1.0                   # Image processing
pytesseract==0.3.10              # Tesseract OCR Python wrapper

# ============================================================================
# Date and Text Processing
# ============================================================================
python-dateutil==2.8.2           # Flexible date parsing
regex==2023.10.3                 # Advanced regex (better than re)
unidecode==1.3.7                 # Text normalization and transliteration

# ============================================================================
# Data Handling
# ============================================================================
pydantic==2.4.2                  # Data validation and settings management
jsonschema==4.19.1               # JSON schema validation

# ============================================================================
# Database
# ============================================================================
sqlalchemy==2.0.22               # SQL toolkit and ORM
alembic==1.12.1                  # Database migrations

# ============================================================================
# Web Framework (Optional - for web interface)
# ============================================================================
fastapi==0.104.1                 # Modern web framework
uvicorn[standard]==0.24.0        # ASGI server
jinja2==3.1.2                    # Template engine
python-multipart==0.0.6          # Form data parsing

# ============================================================================
# Utilities
# ============================================================================
tqdm==4.66.1                     # Progress bars
click==8.1.7                     # CLI creation
python-dotenv==1.0.0             # Environment variable management
pyyaml==6.0.1                    # YAML configuration files

# ============================================================================
# Logging and Monitoring
# ============================================================================
loguru==0.7.2                    # Enhanced logging
prometheus-client==0.19.0        # Metrics collection

# ============================================================================
# Testing (Development)
# ============================================================================
# pytest==7.4.3                  # Testing framework
# pytest-cov==4.1.0              # Coverage reporting
# pytest-mock==3.12.0            # Mocking support
# black==23.11.0                 # Code formatter
# flake8==6.1.0                  # Linting
# mypy==1.7.0                    # Type checking

# ============================================================================
# External System Dependencies (Not via pip)
# ============================================================================
# Tesseract OCR - Install separately:
#   Windows: choco install tesseract
#   Language packs: tesseract-ocr-nld tesseract-ocr-eng
#
# LibreOffice - For DOCX to PDF conversion:
#   Windows: choco install libreoffice
#
# Poppler - For pdf2image:
#   Windows: choco install poppler
#
# ============================================================================

